requirements.txt
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
PyMySQL~=1.1.0
SQLAlchemy~=2.0.31